+++ processor/src/host/mod.rs
@@ -1,17 +1,12 @@
 use alloc::sync::Arc;
 
-use vm_core::{
-    crypto::{hash::RpoDigest, merkle::MerklePath},
-    mast::MastForest,
-    utils::collections::KvMap,
-    AdviceInjector, DebugOptions, Word,
-};
+use vm_core::{crypto::hash::RpoDigest, mast::MastForest, DebugOptions};
 
-use super::{ExecutionError, Felt, ProcessState};
-use crate::MemAdviceProvider;
+use super::{ExecutionError, ProcessState};
+use crate::{KvMap, MemAdviceProvider};
 
 pub(super) mod advice;
-use advice::{AdviceExtractor, AdviceProvider};
+use advice::AdviceProvider;
 
 #[cfg(feature = "std")]
 mod debug;
@@ -22,15 +17,13 @@ pub use mast_forest_store::{MastForestStore, MemMastForestStore};
 // HOST TRAIT
 // ================================================================================================
 
-/// Defines an interface by which the VM can make requests to the host.
+/// Defines an interface by which the VM can interact with the host.
 ///
-/// There are three variants of requests, these can get advice, set advice and invoke the
-/// debug handler. The requests are specified by the [AdviceExtractor], [AdviceInjector] and
-/// [DebugOptions] enums which target the `get_advice`, `set_advice` and `on_debug` methods
-/// respectively. The host is responsible for handling the requests and returning the results to
-/// the VM in the form of [HostResponse]. The host is provided with a reference to the current
-/// state of the VM ([ProcessState]), which it can use to extract the data required to fulfill the
-/// request.
+/// There are four main categories of interactions between the VM and the host:
+/// 1. accessing the advice provider,
+/// 2. getting a library's MAST forest,
+/// 3. handling advice events (which internally mutates the advice provider), and
+/// 4. handling debug and trace events.
 pub trait Host {
     type AdviceProvider: AdviceProvider;
 
@@ -43,20 +36,6 @@ pub trait Host {
     /// Returns a mutable reference to the advice provider.
     fn advice_provider_mut(&mut self) -> &mut Self::AdviceProvider;
 
-    /// Returns the requested advice, specified by [AdviceExtractor], from the host to the VM.
-    fn get_advice(
-        &mut self,
-        process: ProcessState,
-        extractor: AdviceExtractor,
-    ) -> Result<HostResponse, ExecutionError>;
-
-    /// Sets the requested advice, specified by [AdviceInjector], on the host.
-    fn set_advice(
-        &mut self,
-        process: ProcessState,
-        injector: AdviceInjector,
-    ) -> Result<HostResponse, ExecutionError>;
-
     /// Returns MAST forest corresponding to the specified digest, or None if the MAST forest for
     /// this digest could not be found in this [Host].
     fn get_mast_forest(&self, node_digest: &RpoDigest) -> Option<Arc<MastForest>>;
@@ -64,26 +43,8 @@ pub trait Host {
     // PROVIDED METHODS
     // --------------------------------------------------------------------------------------------
 
-    /// Creates a "by reference" host for this instance.
-    ///
-    /// The returned adapter also implements [Host] and will simply mutably borrow this
-    /// instance.
-    fn by_ref(&mut self) -> &mut Self {
-        // this trait follows the same model as
-        // [io::Read](https://doc.rust-lang.org/std/io/trait.Read.html#method.by_ref).
-        //
-        // this approach allows the flexibility to take a host  either as owned or by mutable
-        // reference - both equally compatible with the trait requirements as we implement
-        // `Host` for mutable references of any type that also implements `Host`.
-        self
-    }
-
     /// Handles the event emitted from the VM.
-    fn on_event(
-        &mut self,
-        _process: ProcessState,
-        _event_id: u32,
-    ) -> Result<HostResponse, ExecutionError> {
+    fn on_event(&mut self, _process: ProcessState, _event_id: u32) -> Result<(), ExecutionError> {
         #[cfg(feature = "std")]
         std::println!(
             "Event with id {} emitted at step {} in context {}",
@@ -91,7 +52,7 @@ pub trait Host {
             _process.clk(),
             _process.ctx()
         );
-        Ok(HostResponse::None)
+        Ok(())
     }
 
     /// Handles the debug request from the VM.
@@ -99,18 +60,14 @@ pub trait Host {
         &mut self,
         _process: ProcessState,
         _options: &DebugOptions,
-    ) -> Result<HostResponse, ExecutionError> {
+    ) -> Result<(), ExecutionError> {
         #[cfg(feature = "std")]
         debug::print_debug_info(_process, _options);
-        Ok(HostResponse::None)
+        Ok(())
     }
 
     /// Handles the trace emitted from the VM.
-    fn on_trace(
-        &mut self,
-        _process: ProcessState,
-        _trace_id: u32,
-    ) -> Result<HostResponse, ExecutionError> {
+    fn on_trace(&mut self, _process: ProcessState, _trace_id: u32) -> Result<(), ExecutionError> {
         #[cfg(feature = "std")]
         std::println!(
             "Trace with id {} emitted at step {} in context {}",
@@ -118,7 +75,7 @@ pub trait Host {
             _process.clk(),
             _process.ctx()
         );
-        Ok(HostResponse::None)
+        Ok(())
     }
 
     /// Handles the failure of the assertion instruction.
@@ -128,54 +85,6 @@ pub trait Host {
             err_code,
             err_msg: None,
         }
-    }
-
-    /// Pops an element from the advice stack and returns it.
-    ///
-    /// # Errors
-    /// Returns an error if the advice stack is empty.
-    fn pop_adv_stack(&mut self, process: ProcessState) -> Result<Felt, ExecutionError> {
-        let response = self.get_advice(process, AdviceExtractor::PopStack)?;
-        Ok(response.into())
-    }
-
-    /// Pops a word (4 elements) from the advice stack and returns it.
-    ///
-    /// Note: a word is popped off the stack element-by-element. For example, a `[d, c, b, a, ...]`
-    /// stack (i.e., `d` is at the top of the stack) will yield `[d, c, b, a]`.
-    ///
-    /// # Errors
-    /// Returns an error if the advice stack does not contain a full word.
-    fn pop_adv_stack_word(&mut self, process: ProcessState) -> Result<Word, ExecutionError> {
-        let response = self.get_advice(process, AdviceExtractor::PopStackWord)?;
-        Ok(response.into())
-    }
-
-    /// Pops a double word (8 elements) from the advice stack and returns them.
-    ///
-    /// Note: words are popped off the stack element-by-element. For example, a
-    /// `[h, g, f, e, d, c, b, a, ...]` stack (i.e., `h` is at the top of the stack) will yield
-    /// two words: `[h, g, f,e ], [d, c, b, a]`.
-    ///
-    /// # Errors
-    /// Returns an error if the advice stack does not contain two words.
-    fn pop_adv_stack_dword(&mut self, process: ProcessState) -> Result<[Word; 2], ExecutionError> {
-        let response = self.get_advice(process, AdviceExtractor::PopStackDWord)?;
-        Ok(response.into())
-    }
-
-    /// Returns a path to a node at the specified depth and index in a Merkle tree with the
-    /// specified root.
-    ///
-    /// # Errors
-    /// Returns an error if:
-    /// - A Merkle tree for the specified root cannot be found in this advice provider.
-    /// - The specified depth is either zero or greater than the depth of the Merkle tree identified
-    ///   by the specified root.
-    /// - Path to the node at the specified depth and index is not known to this advice provider.
-    fn get_adv_merkle_path(&mut self, process: ProcessState) -> Result<MerklePath, ExecutionError> {
-        let response = self.get_advice(process, AdviceExtractor::GetMerklePath)?;
-        Ok(response.into())
     }
 }
 
@@ -193,22 +102,6 @@ where
         H::advice_provider_mut(self)
     }
 
-    fn get_advice(
-        &mut self,
-        process: ProcessState,
-        extractor: AdviceExtractor,
-    ) -> Result<HostResponse, ExecutionError> {
-        H::get_advice(self, process, extractor)
-    }
-
-    fn set_advice(
-        &mut self,
-        process: ProcessState,
-        injector: AdviceInjector,
-    ) -> Result<HostResponse, ExecutionError> {
-        H::set_advice(self, process, injector)
-    }
-
     fn get_mast_forest(&self, node_digest: &RpoDigest) -> Option<Arc<MastForest>> {
         H::get_mast_forest(self, node_digest)
     }
@@ -217,77 +110,20 @@ where
         &mut self,
         process: ProcessState,
         options: &DebugOptions,
-    ) -> Result<HostResponse, ExecutionError> {
+    ) -> Result<(), ExecutionError> {
         H::on_debug(self, process, options)
     }
 
-    fn on_event(
-        &mut self,
-        process: ProcessState,
-        event_id: u32,
-    ) -> Result<HostResponse, ExecutionError> {
+    fn on_event(&mut self, process: ProcessState, event_id: u32) -> Result<(), ExecutionError> {
         H::on_event(self, process, event_id)
     }
 
-    fn on_trace(
-        &mut self,
-        process: ProcessState,
-        trace_id: u32,
-    ) -> Result<HostResponse, ExecutionError> {
+    fn on_trace(&mut self, process: ProcessState, trace_id: u32) -> Result<(), ExecutionError> {
         H::on_trace(self, process, trace_id)
     }
 
     fn on_assert_failed(&mut self, process: ProcessState, err_code: u32) -> ExecutionError {
         H::on_assert_failed(self, process, err_code)
-    }
-}
-
-// HOST RESPONSE
-// ================================================================================================
-
-/// Response returned by the host upon successful execution of a [Host] function.
-#[derive(Debug)]
-pub enum HostResponse {
-    MerklePath(MerklePath),
-    DoubleWord([Word; 2]),
-    Word(Word),
-    Element(Felt),
-    None,
-}
-
-impl From<HostResponse> for MerklePath {
-    fn from(response: HostResponse) -> Self {
-        match response {
-            HostResponse::MerklePath(path) => path,
-            _ => panic!("expected MerklePath, but got {:?}", response),
-        }
-    }
-}
-
-impl From<HostResponse> for Word {
-    fn from(response: HostResponse) -> Self {
-        match response {
-            HostResponse::Word(word) => word,
-            _ => panic!("expected Word, but got {:?}", response),
-        }
-    }
-}
-
-impl From<HostResponse> for [Word; 2] {
-    fn from(response: HostResponse) -> Self {
-        match response {
-            HostResponse::DoubleWord(word) => word,
-            _ => panic!("expected DoubleWord, but got {:?}", response),
-        }
-    }
-}
-
-impl From<HostResponse> for Felt {
-    fn from(response: HostResponse) -> Self {
-        match response {
-            HostResponse::Element(element) => element,
-            _ => panic!("expected Element, but got {:?}", response),
-        }
     }
 }
 
@@ -373,22 +209,6 @@ where
 
     fn advice_provider_mut(&mut self) -> &mut Self::AdviceProvider {
         &mut self.adv_provider
-    }
-
-    fn get_advice(
-        &mut self,
-        process: ProcessState,
-        extractor: AdviceExtractor,
-    ) -> Result<HostResponse, ExecutionError> {
-        self.adv_provider.get_advice(process, &extractor)
-    }
-
-    fn set_advice(
-        &mut self,
-        process: ProcessState,
-        injector: AdviceInjector,
-    ) -> Result<HostResponse, ExecutionError> {
-        self.adv_provider.set_advice(process, &injector)
     }
 
     fn get_mast_forest(&self, node_digest: &RpoDigest) -> Option<Arc<MastForest>> {